fix(forest): rung-1 tail — extern-template binding repair + local-class carriability
ROOT 5: the pack drain's body materialization (parseFunction) clears
fd->emit_symbol/declaration_only AFTER the translate-time external-class
binding pass ran, so the freeze snapshotted the extern-template surface
(ostream operator<< family, string ctors) stripped of its libstdc++
bindings — bound consumers emitted restored bodies where live imports the
mangled symbol (_ZNSolsEi). Fix: extract the binding pass into
CirBuilder::bind_external_class_symbols (idempotent, dlsym-verified) and
re-run it at the end of translate_module under pack_recording, so the
frozen state equals a live consumer parse's.

ROOT 6: FUNCTION-LOCAL classes (basic_string::_M_construct's _Guard,
__stoa's _Save_errno/_Range_chk) are un-carriable — their hoisted method
names have no consumer-visible home, and a drain-context pattern build
registers the local class under the REAL owner-qualified key (the
testlocalclassraii hijack, resurfacing because the drain derives with
owner scope). Freeze-layer containment (the parser fix — per-function
local-class identity — is a queued follow-on):
- Program::method_var_is_local_class_hoist / class_is_function_local:
  structural markers (method name not prefixed by its own owner_class).
- forest_arena_record_aggregate skips (and DK_NONE-kills a stale slot
  for) a function-local class; the DK_DEFBODY writer skips local-hoist
  bodies — a restored method-less shell hijacked the consumer's own
  derivation ("no matching constructor for _Guard").
- Pack emission stashes defs (pack_defs) and drops via a post-fixpoint
  CASCADE: a body calling an un-carriable symbol (local-class method, or
  a drop that could NOT revert to DEFBODY) drops too; reverted DEFBODYs
  are safe consumer homes and do not cascade. Every drop is logged.
- tsubst emittable() arm 6: restored forest bodies (has_forest_body,
  collected with the same predicates as translate_module's forest_lazy /
  v20 stages) are definition sources — consumer-side fresh instantiation
  of _M_construct bailed on the CARRIED _M_init_local_buf without it.

Result: bound testsubscript against the drained corpus (2943 evaluated /
348 deferred, 222 logged drops, 696,621 records) runs byte-identical to
live. Reducer pair: tmp/guard_tu.cpp freeze + tmp/guard_use.cpp bind.

+++ src/parser.cpp
@@ -28461,7 +28461,7 @@ TokenBase *TokenCLASS::parse(Program &pgm)
 	      && b.var->name.find(_mti_probe) != std::string::npos )
 		fprintf(stderr, "MTIPROBE eager-body sym=%s file=%s depth=%d\n",
 			b.var->name.c_str(), b.file ? b.file : "(none)",
-			pgm.fn_template_instantiation_depth);
+			(int)pgm.fn_template_instantiation_depth);
 	    pgm.parse_deferred_function_body(b);
 	}
     }
